Temporarily fix import/export issues resulting
These will have to be reconsidered at a latter step; not sure why
duplication of exports are now required (that is, default vs non default;
why was it not required before?). (+1 squashed commit)
Squashed commits:
[0f0f13f] Temporarily fix import/export issues resulting

These will have to be reconsidered at a latter step; not sure why
duplication of exports are now required (that is, default vs non default;
why was it not required before?).